Improved setup_pips.sh:
  Can resume pips installation after some interruption. Useful when fixing
  some stuff during installation.

  wget can work even with my .wgetrc :-) (that forces to create a full
  directory path matching the URL)

Added warning messages about all the error messages displayed during the
bootstrap of the .h files in phase1 that are normal.

Now directories can be installed in build_etc. It was missing for
the complexity analysis library.
By the way, the command "install" can only with plain files, not
directories...

Added directory definitions about X11 and xview for wpips.

Cleaned up makefiles to have or not wpips in fpips.


git-svn-id: http://svn.cri.ensmp.fr/svn/nlpmake/trunk@825 0884f79e-42f2-0310-bc61-a3b21e0d3bdb