DomainAddRays: allow addition of zero rays
A Matrix with zero rows doesn't have a valid p pointer.
The original code would dereference this pointer anyway, resulting
in a Segmentation Fault.
Instead, we now return a copy of the original domain if the number
or rows in Ray is zero.
We also return a copy when Ray itself is zero, rather than the
original domain itself, since the caller expects the result to
be a newly allocated domain.