The HTML generation is back! Added or modified some Makefiles for HTML versions of dev_guide, pipsmake, properties, tpips-user-manual.
Is it interesting to have all the NewGen descriptions in HTML too?

What about the HTML files of the web site?
I think it is time to think again on the web site of the PIPS project that
uses a 10-year old technology... :-)

What about a CMS as front-end with some parts generated and updated by a
Makefile with a sensible scp or even a WebDAV access on a Plone site for
example?


git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@12472 32a0c1e6-9308-0410-80dd-f9d2257a62c7