Added a new INTERNALIZE_PARALLEL_CODE transformation to coerce a parallel code into plain one for PIPS, that is something we can work on again later, with optimizations, other kind of parallelization and so on. I love writing so simple transformations. :-)
Note that right now, a parallel code displayed later as PRINTED_FILE will
not be displayed with parallel attributes since they are not displayed by
this sequential pretty printer.

For PIPS developers: what about adding some user properties to enable the
display of parallel stuffs even in the sequential pretty printer if they
exist?

Another glitch: if we add a new phase, we still need to manually rebuild
Libs/pipsmake. But this has been discussed far in the past, even if I
can't remember the religious reason behind it... :-)


git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@12480 32a0c1e6-9308-0410-80dd-f9d2257a62c7