matrix_addon.c: fix typo in mpolyhedron_{in,de}flate
The original code would ignore one of the coefficients it was
supposed to take into account and instead use the "status" value
by mistake.

For example, ehrhart_upper_bound < Test/ehrhart/e0.in would give

         P  >= 0
          1 >= 0

( 1/4 * P^2 + 1 * P + 3/4 )

which is incorrect for (at least) P = 0, where the value is 1 (>3/4).
Now, the result is

         P  >= 0
          1 >= 0

( 1/4 * P^2 + 3/2 * P + 9/4 )