Param_Polyhedron: keep track of constraints saturated by parametric vertices
Some applications, most notably the barvinok library, need
to know which constraints are saturated by each vertex.
This information is (mostly) available in scan_m_face,
while storing it shouldn't require too much memory (1 bit per
constraint, per vertex).

The only extra computation added by this patch is that we now
also need to check whether any of the constraints not explicitly
selected also saturates the vertex.

To make sure that a user of a Param_Polyhedron knows which
constraints are referred to, we now also keep a copy of them
in the Param_Polyhedron structure.