align_context: check that every polyhedron in the domain has the same dimension
Return NULL if they do not.
Also fix the return value in case of a context dimension that is too big
and in case the dimension is the same as the required dimension,
where we need to copy the whole domain instead of just the first
polyhedron in the domain.