pushed revisions 12852:12908 from branch http://svn.cri.ensmp.fr/svn/pips/branches/keryell/test to master http://svn.cri.ensmp.fr/svn/pips/trunk
If that works, Fabien's svn_branch.sh is just great! :-)

New version of the C parser around comment parsing.

Fixed the code to pass without core-dumping on the hologram code from TB
fresnelslowP_pips.c (to be added to the validation...).
It was my motivation to fix the parser. Now push and pull are symmetric on
the stack. :-)

Simplified the comment C parsing code that was a little cumbersome.

Simplified and fixed the C comment prettyprinter that was relying on '\n'
everywhere.

Improved the C comment parser to accept line spaces as comments to improve
the user code rendering.

Reorganized the code dealing with comments in the C parser and done some
cleaning.
Added code to deal with line spacing.
Cleaned-up the C comment prettyprinter to be more in the sentence spirit.

Moved some generic code to ri-util form parsers and
transformations/array_resizing_top_down.c.

TODO: there are huge parts of PIPS that should be moved to ri-util... :-(

I've tested the validation on my Debian unstable x86_64 but not accepted
anything (since the CRI is working on the validation right now). Globally,
that looks better, even if not perfect. The issue is that it is difficult
without using contextual start conditions that are not used in the lexer
right now. :-(

Here are the problems I noticed in the C_syntax validation:

  Funny output: C_syntax/array.result/out

  Empty output: C_syntax/files.result/out C_syntax/scope01.result/out
  C_syntax/scope02.result/out

  Strange comments: C_syntax/goto.result/out

  Wrong code: C_syntax/unstructured01.result/out


git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@12909 32a0c1e6-9308-0410-80dd-f9d2257a62c7