Fix infinite loop in ZDomainDifference
The original loop was replaced by its first iteration in
162979ddc7eee2196df18d5f8221c77a661e1f86 (SplitZPolyhedron function
introduced.), without any explanation.  Commit
1a019024f1b3d0eceb1a4d678f3fc568a34397f2 (ZDomainDifference: add
back commented out code) attempted to reinstate the loop, but failed
to drop the initialization, resulting in an infinite loop.

[Bug reported by Armin Groesslinger, Univ. Passau]

Note, however, that even with this inifite loop fixed,
the original computation now finishes, but results in
the wrong solution.  See
http://lipforge.ens-lyon.fr/tracker/index.php?func=detail&aid=3171&group_id=52&atid=296
for more information.