polyparam.c: find_m_faces: collect rays of unbounded parametric polyhedra
If there were lines in the resulting polyhedron, there was a chance that
parametric vertices were not computed correctly [reported by Salvatore
Ruggieri]. Asked a request by Salvatore in the same time: as I needed to
compute the lines and unidirectional rays, I added them in the resulting
structure, as a PolyLib-formated matrix of rays. Structure change in
types.h. Added some tests in Test/pp to check correctness (some of them
are from Salvatore, some by Vincent).

Tested-by: Salvatore Ruggieri