Factorized itoa out, since it is now exported from NewGen genC. Fix some 64-bit issues and other warnings.
How to spend 2 days on another little typing warning?
Just begin with control/unreachable.c that happens to work by side effect
by mixing statements and controls in the same mapping.
Try to fix the code that does no longer work... Then rewrite some parts of
the recursion... :-(

I also improved it to deal with continuation status on unstructured with
unreachable exit.

Not yet validated because of a bug in control/bourdoncle.c ... :-/


git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@13315 32a0c1e6-9308-0410-80dd-f9d2257a62c7