Some interesting examples.
I've put NS_SIMPLE_IO.tpips because I first thought that the code was not
correctly parallelized but is the PIPS beauty:
the loop with IO is parallelized because it is not... executed. :-) If it
is executed (I've copied the same loop-nest in the other if-branch), it is
not parallelized. :-)

parallel_to_sequential.tpips is an example to show that a parallelized_code
resource can be worked on later with other passes. It needs some deeper
verification anyway.


git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@409 67b5017d-e62d-0410-a871-f72b1c62810f