Working P4A Cuda version. Corrected many bugs.
The more vicious one was that the image reading and writing functions
I stole in the Holotetrix Fresnel application was wrong... It mixed up
line and columns, but since it was only used on square symmetric
images, it never happened... :-/

Cuda performances are poor since it is a memory bound algorithm...

To be implemented by some PhD students: aggressive allocation in
shared memory when all the images can fit (with nice overlap stuff
dedicated to Fabien). Shared memory could be be used as a cache anyway
(a sliding window in the image processing jargon), to spare some
memory bandwidth anyway.

Added an HPC Project logo to be used as a demo image. :-)


git-svn-id: svn+ssh://svn.freia.enstb.org/subversion/freia@247 576080ae-5d29-4e32-8499-467c2d9a9582