Avoid crashing on unstructured tests in if_conversion_init
if_conversion_init should be improved to deal with unstructured stuff.
I've added some code to skip unstructured test. So, the if conversion is not done in this case... :-(
I guess it is similar to what is already done in FSM generation in Libs/phrase.
Trying to factor some code out from fsmize_statement()?
So this issue is not related to the controlizer or the restructurer... :-)

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@16026 32a0c1e6-9308-0410-80dd-f9d2257a62c7