Added the lacking declaration of the function before its call by the oultiner.
That strange, it used to work in October.
By looking in the validation, the code generated was invalid, but
accepted. :-/
I guess we could outline some code that generates C99 code because the
declaration is not add at the beginning of a sequence.
Should be its own directory, with inlining and outlining in different files.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@16052 32a0c1e6-9308-0410-80dd-f9d2257a62c7