p4a_recover_includes use now a real lexer (PLY).
Switched to the lexer from PLY (Python Lex Yacc) pointed by Laurent Daverio to correctly match { }.
It now parses via PIPS all the .h of the GNU libc from the documentation (and it shows some bugs in the libc...).
Can now optionally direct the output to another file.
Updated to the PIPS file header that is now 3 lines instead of 7.