Added a lacking initialization that made the validation to flick.
How is this kind of bug still possible?
Added also a description for open_module_if_unique(). It took me some time
to guess first what was the aim of it and then what was the expected
behaviour...

Why having comments in PIPS since it is so trivial to understand? :-(

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@16207 32a0c1e6-9308-0410-80dd-f9d2257a62c7