Updated Jacobi example to suit recent PIPS & PyPS developments
Heavy usage of PyPS.
Now the GPU-ification is split in 2 parts, with a loop nest annotation in the middle.
Added reprivatization and reparallelization since it is lost in the outliner... parser. :-(
Now the Makefile use the .py instead of .tpips
The post-processor needs to be updated to generate correct code.