Added a property to deal with parallel loops in GPU_LOOP_NEST_ANNOTATE
By setting the GPU_LOOP_NEST_ANNOTATE_PARALLEL property to true, only the
top-level parallel loop nests are annotated.

For Serge: I cannot set this property in PyPS, so I set it to TRUE by default.
Could you see why?

For Béatrice: the consequence for this is that the validation is changed.
Could you add some other validation tests for this property as TRUE such as:

do par
  do seq

do par
  do par

do seq
  do par
(no annotation on this)

Thank you.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@17196 32a0c1e6-9308-0410-80dd-f9d2257a62c7