p4a_setup missed .cu file.
I've already fixed this issues recently but it may have disappeared in
some merge.

Added also a copy in src/dev/p4a_setup-rk so that this version can
co-exist with Grégoire version.

Added a comment in the release notes on the fact that PLY is not installed
to be usable by default in Debian with Python 2.6 and 3.1. This has side
effects on ipyps that breaks down...

TODO : bug reporting to Debian to have python-ply usable with all Python
versions.