Do not reset locale before launching PIPS.
p4a was failing on wcdma.c example. In fact it was due to a French accent
not accepted by the #include recovering.

I first thought to PLY not compatible with the new Python 3.1 encoding
stuff, but indeed it was that the locale was reset to plain old C with
ASCII encoding... :-/