Improved the prettyprinter to output #pragma and label in C
OpenMP wants pragma between a label and a statement in C, not before the label.
Note that this breaks some Fortran a priori in the pretty-printer with
text.tex...

In some case extensions were not prettyprinted.

Improved Doxygen comments and some cleaning.

This corrects a bug with OpenMP generation on wcdma.c from SCALOPES project.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@17574 32a0c1e6-9308-0410-80dd-f9d2257a62c7