Added examples showing how to dynamically inject Python code to change p4a behaviour
Changed some class definitions to new style to have super() working.
Added 2 examples showing how to change a method or a class inside p4a.
Improved the documentation of the p4a architecture describing these.
Examples could be simpler in Python 3...

Could serve as example for an article on PyPS. :-)