Some warning fixes.
It is important to declare functions before using it.
The compiler only issues a warning and resulting code may be ok on local
architecture, but the compiler assumes the declaration:
int fun();
for undeclared functions.
If the function returns a pointer, there is an implicit conversion from
int to pointer,
It ends in an error if sizeof(int) < sizeof(void*)

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18160 32a0c1e6-9308-0410-80dd-f9d2257a62c7