Improved documentation of the RI.
Why using complex LaTeX infrastructure if there is NO documentation on
some types? :-(

There are many things in ri_C.tex that should be copied into
ri.tex... ri_C.tex is rather a discussion document on what is done, on
what is indeed not done in the RI... So if something is really in the RI,
it should be really described in ri.tex!

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18169 32a0c1e6-9308-0410-80dd-f9d2257a62c7