ICM failed badly, and cause this test to abort.
As ICM pass has been added last week, I think that removing it now won't change the initial mind of the test : 

"This test is useful to make sure that preconditions are propagated efficiently within loops with very long linear bodies. The performance bang was due to a lack of normalization"

Feel free to add again ICM, but if it still fails, you have to move the test to Scilab-Bugs.





git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@2624 67b5017d-e62d-0410-a871-f72b1c62810f