Rationalized DMA runtime functions in kernel_load_store and isolate_statement
Now we could have a C99 implementation of the runtime.

Now the scalar transfer case is special (so it could be implemented in
hardware FIFO...).

Keep passing 1 useless parameter for symmetry reason, because of C/Fortran
array subtlety to think further about later...

Improved documentation.
ADDED A LOT OF COMMENTS... (<<< this is a subliminal advertisement about what we should have in PIPS :-/ )

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18320 32a0c1e6-9308-0410-80dd-f9d2257a62c7