Renamed single_assignment as scalar_renaming.
The algorithm is completely re-written too, and is now correct (as in all
validation files are correct, not as proved).

Some files should have been renamed, but they were not.
Also, there is no good reason to keep this transfo in sac, but history
will prevail.

Arf, I lost one day of phd redaction :-(

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18331 32a0c1e6-9308-0410-80dd-f9d2257a62c7