Fix a bug in replace_entity:
when you replace an entity, all parent expression_normalized become invalid
because they still refer to the former value.

e.g.
      (1)
       |
       +
      / \
     /   \
    (2) (3)
     |   |
     5   h

normalized (2) -> 5
normalized (3) -> h
normalized (1) -> h + 5

after replacement of h by h'
you got
      (1)
       |
       +
      / \
     /   \
    (2) (3)
     |   |
     5   h'

normalized (2) -> 5
normalized (3) -> h
normalized (1) -> h + 5

this is incoherent and leads to errors in following computations.

replace_entity now take care of this for you,
but any time an expression is changed, we have this risk pending.

Why ? because the normalized field of an expression by passes pipsmake.
It should be either a resource, or a table computed when one enters a
new phase ...

Just so you know, I've lost two days because of this bug, but there is a
nice side effect: polynomes are now compatible with pips expressions :)

BTW, when we call NORMALIZE_EXPRESSION, maybe we should first try to
simplify it using simplify_expression, just in case the expression can
be reduces to a simpler (linear) expression ...

^^! that's a huge commit log

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18354 32a0c1e6-9308-0410-80dd-f9d2257a62c7