Introduce some typedef for PIPS
P4A_accel_kernel_wrapper & P4A_accel_kernel are intended to be the return type of kernel_wrapper & kernel functions.
A #define will later replace these with __global__ or __device__ if we want to compile it on cuda.
These typedef will be used in PIPS to replace the default (void) return type, so that the prettyprinter output include these return type and thus avoid us to do some postprocessing.