Some more update for kernel_load_store transfer generation.
It uses regions to generate DMA transfers.
the may/must field of the data-out is used to choose the amount of data
to copy-in,
makeing the propery FORCE_LOAD obsolete.

As the full engine is now based on regions, cases may arise where
regions are not accurate enough.
In that case, code is generated based on the definition regions.
If they can not be represented as a linear expression, partial_eval is
tried to simplify their expression again.

I tried hard to validate this transformation against pips' validation
and to make the transformation as accurate as possible (even if I am not
using in/out regions),
so any misplaced remarks from RK will not be appreciated :)

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18548 32a0c1e6-9308-0410-80dd-f9d2257a62c7