yet another isolate_statement enhancement.
Asked by RK.
The aim is to extand isolate_statement to array of structs / struct of
arrays / struct of array of structs of ...

This is a first approximate, more work is needed to hndle corner cases
accurately.

Anyways here is the scheme:
each region is converted to its rectangular hull,
fields are pruned when not at the head of the access path
and the region approximation becomes may, which leads to extra dma.

A particular case is handled very inefficiently:
if all fields of a struct are accessed, we still generate a may
approximation and extra dmas.

I have added comment in the source code about how to do this, after a
phone talk with BC.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18598 32a0c1e6-9308-0410-80dd-f9d2257a62c7