Remove the dependency from maxima in accelerator util.
It now relies on the polynomial implementation from linear I extended
with the crude `polynome_solve'.

based on this, I was capable of implementing computation_intensity
detection for pips.
the basic idea is to enumerate regions of a statement and add
resulting polynomials, then substract complexity and consider the
behavior of the resulting polynomial toward the infinity.
A positive value means the loop is suitable for offloading /
parallelisation whatever.

More on this (but not that much more) in pipsmake-rc and in the thesis
:p

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18657 32a0c1e6-9308-0410-80dd-f9d2257a62c7