Add memory_optimization_runtime.cpp
The runtime for memory optimization keep track of associations between memory area in host soace  and corresponding memory area in the accelerator memory space.

This is a very basic runtime at that time, it's only some wrapper around c++ "map" ; we'll have to check in the future that we have enough space for allocation and probably have to free space on a kind of LRU strategy.