Nicer handling of comments in PIPS.
autogeneration of the comments guard if none is found.
Now it is no longer necessary to prefix statement comments by a C or //
(or ! or /* or c or # ...) the prettyprinter takes care of this for you.
Multiline support included !

This fixes a few bugs in the validation (challenge: can you spot them ?)
And a property can be used to turn this off, some people use comments to
add #includes in hpfc :p

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18744 32a0c1e6-9308-0410-80dd-f9d2257a62c7