Updated Gpu/jacobi01.py
I guess it was not correct for a while...
I spent a lot of time on a typo that make some properties disappearing because of a PyPS bug (still to fix):
w.props.CONSTANT_PATH_EFFECTS = False,
(note the ",") create a property with a list value "(False,)" that is ignored.
Removed spurious trailing spaces.

git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@2975 67b5017d-e62d-0410-a871-f72b1c62810f