First attempt to have user error and warning to work in property stuff.
I spent days tracking bugs before because there was wrong property names
used in some validation cases, hiding bugs because there was a discreet
message instead of an usual abort when ABORT_ON_USER_ERROR is set. :-(

There was a special implementation of pips error and warning to avoid
recursion when the error handler itself fails in properties.l because it
needs some properties to select is behaviour, leading to infinite error
recursion...
Now it should work better.

For example, even in property stuff, ABORT_ON_USER_ERROR is now used to
select the behaviour.

Now the redefinition of a property is just a warning. I cannot see why it
was an error before...

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18813 32a0c1e6-9308-0410-80dd-f9d2257a62c7