Add complex basic compatibility layer for p4a.
In the form of the p4a_fcomplex type.
It uses c++ operator overloading to do the trick, where cuda's implementation is function based.
The implementation is more a proff of concept than a complete implementation ...

Also adds a fallback for PIPS's MIN MAX.
In the form of #define.
I am unsure p4a_accel.h is the good place to do this.
But it is always included, so it is convenient ...