Clean up Jacobi demo
Removed old jacobi.py and jacobi.tpips since it is out of sync and p4a is
the native way to go now.
Improved Makefile to cope with Hyantes demo (Makefile to factorize out by
the way).
Added communication optimization entries in the Makefile.
Seems to fail with current version with an infinite loop.