The C99 parser for PIPS dealing with declarations anywhere and for loops
After a day in François office, things are always brighter. :-)
Interestingly it simplifies the parser since there is no longer a symmetry
breaking between declarations first and other statements afterwards.
Keep more comments from the user sources in for(...) clauses.
Improved comments.
Some code cleaning in ri-util and c_syntax.

TODO: optimizing RI when generating for(int i;...) to keep int i in the
for statement and not in a block around.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18939 32a0c1e6-9308-0410-80dd-f9d2257a62c7