Improved comment capture inside for clause by the C parser
Try to save a max of comments. The issue is that opt_expression's can
reset the comments if there is a comma_expression in them. :-( So at least
preserve the commemts before the for.

Issue trigered by several examples such as
validation/Semantics-New/do01.tpips

But I think now that, in a *source-to-source* compiler, comments should appear
*explicitly* in the parser syntax and not be dealt by some side effects in
the lexer as now in PIPS with some stacks and so on.
By the way, this is how I deal with #pragma in the parser...

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18940 32a0c1e6-9308-0410-80dd-f9d2257a62c7