Added a test case for C99 declaration after a non declaration statement.
OK, there is only one validation test on it.
I rely on everybody to try and to provide validation cases or bug reports.
Onil and Janice at HPC Project should provide some.
Time to try real applications and real benchmarks. :-)
I guess the main issues will not be in the parser but in the other PIPS
phases...

git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@3118 67b5017d-e62d-0410-a871-f72b1c62810f