change isolate_statement's handling of prefix. This impacts also kernel_load_store since it passes KERNEL_LOAD_STORE_VAR_PREFIX property to do_isolate_statement. Hope this does not impact too much par4all :-(.
Before this modification, the prefix was added before the
entity_local_name of the original variable. Hence, for variable named
"a" and prefix "P4A_" the generated name was "main:0'P4A_var_0'a0" and
prettyprinted as "a0".

Now the generated name is "main:0'P4A_a0" which is prettyprinted as
"P4A_var_a0". Generated variables can then be detected in generated files.




git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@18961 32a0c1e6-9308-0410-80dd-f9d2257a62c7