Improved C99 for-loops with embedded index declaration
Now can deal with scope declaration to avoid conflicts names on index
variables.

Began to think to a C_PARSER_GENERATE_COMPACT_C99_FOR_LOOP_DECLARATION
version suggested by François Irigoin, with loops holding declaration
entities and been considered as declaration statements too...
Not finished.
It is complex because of for-to-while and for-to-do loop conversions in
PIPS and other details.

So, right now, I consider it is enough to cope with C99 applications, even
if the pretty-print output is less nice.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@19013 32a0c1e6-9308-0410-80dd-f9d2257a62c7