Factorized C/Fortran loop normalization
Please write concise code... Because it is shown on slides!

Anyway, I'm not sure this code is correct.
Loop normalization for GPU is to have compatible loop iteration space with
hardware iteration space and I cannot see why it should depend from the
underlying language...
Has this code been tested?