Revert filtering chains using scope, as stated in pipsmake :
"It is possible to mask effects on local variables in loop bodies. This
is dangerous with current version of Allen \& Kennedy which assumes that
all the edges are present"

And : 

"These last two properties are not consistent with PIPS current development
(1995/96). It is assumed that {\em all} dependence arcs are present in the
dependence graph. Phases using the latter should be able to filter out
irrelevant arcs, e.g. pertaining to privatized variables."




git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@19229 32a0c1e6-9308-0410-80dd-f9d2257a62c7