improve genericity of binary operators (new generic_binary_op) and propagate the difference to callers
There is a new generic engine to replace the old one, as new specific
engines were sufficiently similar to allow merging them.

Binary operators that had not been previously extended to handle
abstract locations (such as inf_difference, "entities_intersection",
...)  now rely on the new generic engine.

The generic engine is called with two arguments: a generic function
specific for a kind of operation (union, intersection,...) and a
function to compute the result for two concrete effects in a
particular representation... Hence the way abstract locations are
handled is clearly separated from the list of effects binary engine
and from the individual concrete effect operators.

In particular, this should allow to change the merging policy more easily if need be, and
it will ease maintenance.

Also, some comments added...




git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@19871 32a0c1e6-9308-0410-80dd-f9d2257a62c7