Updated changelog to Par4All 1.2
This version targets mainly the Wild Cruncher, a parallelizing environment
from HPC Project for Scilab programs. Par4All is used to parallelize the
output of the Scilab-to-C compiler from HPC Project.

Please, do not decide to do big work right now in Par4All or PIPS since we
are just in the finalization of Wild Cruncher... :-)