fix a very small typo in entities_maymust_conflict_p which led to a huge bug! (formal entities and all_locations entities were sometimes considered as non-conflicting, depending on the order in which they were passed to the function).
This brings to light a bug in inlining, which shows up when there are
anywhere effects. The current solution is to set CONSTANT_PATH_EFFECTS
to FALSE until this is fixed up. I'll write a ticket about this.

Also continue some cleaning and adding comments in
references_may_conflict_p for further improvements.



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20002 32a0c1e6-9308-0410-80dd-f9d2257a62c7