Change P4A Cuda Accel initialization
P4A_init_accel is now defined as a function (seems cleaner).
This initializer is always called in the main as usually,
now it read from the environment P4A_MAX_TPB which is (if
an integer is found) assigned to p4a_max_threads_per_block.
This new variable replace P4A_CUDA_THREAD_MAX which was
intended to fix the maximum number of threads per block
at compile time. It's still applicable if P4A_MAX_TPB is
not found (or is invalid) in the environment.