Fix ylwrap so that absolute file names are working back
I'm not very happy with all this ylwrap kludge to deal with old lex/yacc
instead of doing all with flex/bison in a native way.  Compatibility with
the past...

May be the modification from Serge in the previous commit could be rethought
instead of applying a correction on it with this commit?

git-svn-id: http://svn.cri.ensmp.fr/svn/nlpmake/trunk@1143 0884f79e-42f2-0310-bc61-a3b21e0d3bdb