Fix a bug in entity_scalar_p.
A pointer was considered a scalar, although the doc of the function
states that if the function returns true, the entity can not be
dereferenced.

Can you imagine that such an important function was buggy ;-)

Several validation files proved to be false because of this bug, and I
changed a few tests in FC's code (and in mine) to reflect the changes.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20295 32a0c1e6-9308-0410-80dd-f9d2257a62c7