Benchmarks : fix precision in the generate_speedup script
GNU bc is used to compute the geometric mean. "scale" was used to
round the result so that it would be printed nicely in the graph.
The issue was that bc truncate the intermediate results according
to the given scale. This might lead to a noticable imprecision in
some cases.