Huge change in chains computation : do not keep and duplicate intermediate DEF and REF set for each statement.
Previously DEF_IN, DEF_OUT, REF_IN, and REF_OUT were computed *and stored* for each statement. This could 
consume a lot of memory. Now there's a single set for DEF and another single set for REF. For each statement, 
the sets are computed and the conflict in the dependence graph are directly added. The sets are then used 
for the next statement.



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20427 32a0c1e6-9308-0410-80dd-f9d2257a62c7