fix std file entities testing.
The problems were due to the fact that the function could be called
before any of these entities have been created. The chosen solution is
to force the creation of stdin, stdout and stderr entities as soon as
the testing functions are called. This may however prevent testing
that stdio.h is included when needed.



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20471 32a0c1e6-9308-0410-80dd-f9d2257a62c7