Fix GPU_IFY and new compilation unit properties.
It's no longer needed to hack and manually merge files after
save to get kernel and launcher in the same file.

I fixed OpenCL, but a similar process seems to be used for
CUDA-Fortran and thus it should also be adapted accordingly.

A lot (lot, lot, lot, ...) of refactoring is needed :-(


Former-commit-id: 439f3f9b8809003e504f3fab1a0e3ec84a7cffc2