Fix a disruptive behavior in the preprocessor.
In some version of gcc (at least 4.6), system header files may contain
inline functions. These functions are processed by pips as regular
function and thus added to the module list.

This is very disruptive, as it means that the module list depends on the
system configuration, and thus the validation becomes very
system-dependent (extra modules are printed by %ALLFUNC depending on the
gcc version).

This commit somehow fixes this behavior by tracking functions declared
in a header file and not adding them to the module list. It is not a
very good fix because it fails in the presence of #included .c file or
other unusual (but valid) patterns. At least in that case the previous
behavior is preserved.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20524 32a0c1e6-9308-0410-80dd-f9d2257a62c7