A better solution for opengl display.
The problem is actually in the use of glutMainLoop() function
which never returns. GLUT was designed around the idea of
an event-driven application, with the exit method being
captured through an input event callback routine,
such as a GLUT menu or keyboard callback handler (from :
http://www.opengl.org/resources/faq/technical/glut.htm and
http://www.sjbaker.org/steve/software/glut_hack.html)

Using freeglut, this function is supposed to return to the calling function
if the option GLUT_ACTION_ON_WINDOW_CLOSE is used. It does not yet work well,
so the window has to be closed at the end of the execution to allow
the execution of the next demo (in case of the chain of demo).


Former-commit-id: e249c9b42e1486aedc74e4e176510415e1f57b79