Add new RELEASE-NOTES file and put the old one in RELEASE-NOTES-Archives
The new RELEASE-NOTES using a different way to organise the file.
Please check if the items that I put in "Additions" are all right in their place.
Some items from the old RELEASE-NOTES have also been put in the knowing issues.
And if you have input for some additions, changes, removals and fixes (fixed bugs)
items please do not hesitate to add them.


Former-commit-id: dec1170fb35a08a6b6736c03b048c4154a71e788