Replace a redundant remote copy by a symbolic link during package publishing
Since a .deb package is both in the sorted by date organization and in the
repository, use a link from the later one to former one.

Save around a 100 MB copy per publishing each type and 100 MB (it is great
for my ADSL connection :-) ) of remote storage total.

Add comments.
Simplify code.


Former-commit-id: 066bd9b5d8d4e1574357c2b36ca7d940b2a57488