static variables to store entities from standard libraries: they are now resetted before and after a phase is applied.
This is necessary because a phase may invoke get_stdin_entity on a
module from a computational unit which does not include stdio.h (and
then set the static variables to entity_undefined), and then analyze
other modules which include stdio.h.

I also created new functions to hold the different
initialisations/resets done before and after performing a phase on a
module.



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20654 32a0c1e6-9308-0410-80dd-f9d2257a62c7