Use python2.6 explicitly for modern Ubuntu and Debian
Since we distribute a compiled version for a given Python version, fix a
version number so that we have the right Python on the machine.

This problem happens with the last Ubuntu 11.10 and Debian versions where the
default Python versions are not 2.6.
The constrain is to be both compatible Ubuntu LTS and last Ubuntu versions.

This issue was spotted by Cedric Nugteren from Eindhoven University of
Technology.


Former-commit-id: 6d4b99573ca47975ff608310e84ac64b9516fc98