In effects of loop: fix for simple effects.
Also, enhancements for rw effects of loop due to a modification of
simple_effects_union_over_range. When the loop bounds are numerically
known, a test is performed to know if the loop is executed at least
once.

Add a new function in loop.c to know when a range contains at leat one
point.



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@20731 32a0c1e6-9308-0410-80dd-f9d2257a62c7