[P4A CORE] No longer inline kernel's content.
This is a huge change, as it implies that outlining in a new compilation
unit brings all the sub callgraph, associated types etc to the new
compilation unit. It also changes the way we call gpu_ify, because we
cannot gpu_ify the subcallgraph of an already gpu_ified function.

This change is motivated by the fact that unfolding sometimes leads to
complex code that cannot be analyzed by PIPS. It was the cas for the
Chirp code from thales.


Former-commit-id: 81045e9349bcbcb65eda11072e683e42570e2436