Apply flatten_code before parallelization since it break it now
Among the little change one interesting case shows that 
loop fusion shouldn't be applied immediately after parallelization.

For instance:

forall()
  forseq() 
    ...
  forall()
    ...
The inner loops won't be fused because one is // and not the other.
However if we chose generate a coarse_grain code, we will have:

forall()
  forseq() 
    ...
  forseq()
    ...

And then the inner loops can be fused!




git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@5448 67b5017d-e62d-0410-a871-f72b1c62810f