[points-to] remove compiler warning that prevents compiling with -Werror
Error message was:

../../../../src/Libs/alias-classes/expression.c: In function ‘struct_assignment_to_points_to’:
../../../../src/Libs/alias-classes/expression.c:1418: error: suggest parentheses around ‘&&’ within ‘||’

I placed parentheses as seemed the soundest to me, please check!

It would be a good idea to check there is no compilation warning
before commiting. Apart from the fact that several of us
systematically compile with -Werror, and that this prevents us from
compiling, those warning are often hints that something may go wrong
during the execution...



git-svn-id: https://scm.cri.ensmp.fr/svn/pips/trunk@21308 32a0c1e6-9308-0410-80dd-f9d2257a62c7