[Privatization] attempt at having rules requiring privatization independent of privatization rule
Introduce a new resource (MODULE.privatized) produced by every
privatization phase, and which is required by some phases intead of
using !MODULE.privatize_module.

This cannot be used for kernelize which requires that privatize_module
and then coarse_grain_parallelization be applied in this particular
order.

Documentation added in pipsmake-rc to explain why it is sometimes
tricky to privatize global variables.



git-svn-id: https://scm.cri.ensmp.fr/svn/pips/trunk@21338 32a0c1e6-9308-0410-80dd-f9d2257a62c7