Improve p4a_recover_includes to remove parasitic static functions from the libc
In gcc 4.8 some static functions of the libc are parsed by PIPS and
remains afterwards in the user code, conflicting with the libc.
I first tried to change Par4All and PIPS/unsplit() to skip the pretty-printing
of some functions but it became quickly unreasonable to do so.
So I ended up to add a post-cleaning phase in p4a_recover_includes.


Former-commit-id: c245253f97923df1993ca0396569ca2bf5fd9044